# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ConnectionTestResponse": {
        "properties": {
          "attempt": {
            "title": "Attempt",
            "type": "integer"
          },
          "checks": {
            "items": {
              "$ref": "#/components/schemas/DiagnosticResponse"
            },
            "title": "Checks",
            "type": "array"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "data_source_id": {
            "format": "uuid",
            "title": "Data Source Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "overall_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overall Code"
          },
          "poll_url": {
            "title": "Poll Url",
            "type": "string"
          },
          "queued_at": {
            "format": "date-time",
            "title": "Queued At",
            "type": "string"
          },
          "source_version": {
            "title": "Source Version",
            "type": "integer"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "stale"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "data_source_id",
          "source_version",
          "status",
          "checks",
          "overall_code",
          "attempt",
          "queued_at",
          "started_at",
          "completed_at",
          "poll_url"
        ],
        "title": "ConnectionTestResponse",
        "type": "object"
      },
...
      "DiagnosticResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "duration_ms": {
            "title": "Duration Ms",
            "type": "integer"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "remediation_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remediation Hint"
          },
          "status": {
            "enum": [
              "pass",
              "fail",
              "skipped"
            ],
            "title": "Status",
            "type": "string"
          },
          "type": {
            "enum": [
              "network",
              "tls",
              "authentication",
              "authorization",
              "metadata_access",
              "latency"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "status",
          "code",
          "message",
          "duration_ms"
        ],
        "title": "DiagnosticResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Every capability is exposed through this versioned API; the web application is one client among potential others (principle 8). No endpoint accepts SQL, a formula, or a field path from a client.",
    "summary": "Phase 1A — platform skeleton. No business intelligence functionality.",
    "title": "TriVera Executive Intelligence Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/connection-tests/{job_id}": {
      "get": {
        "operationId": "get_test_v1_connection_tests__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionTestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Test",
        "tags": [
          "connection-tests"
        ]
      }
    },
...
    "/v1/data-sources/{source_id}/connection-tests/latest": {
      "get": {
        "operationId": "get_latest_test_v1_data_sources__source_id__connection_tests_latest_get",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Source Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionTestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Latest Test",
        "tags": [
          "connection-tests"
        ]
      }
    },
    "/v1/data-sources/{source_id}/test": {
      "post": {
        "operationId": "request_test_v1_data_sources__source_id__test_post",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Source Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionTestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request Test",
        "tags": [
          "connection-tests"
        ]
      }
    },
...
  }
}